+++ examples/codec_test/spanalysis.cpp
@@ -49,7 +49,7 @@ void usage(char *argv) {
 //-----------------------------------------------------------------------------
 // Set parameters from command line options
 //-----------------------------------------------------------------------------
-int SetOption(int argc, char **argv, int *fft_size, double *q1,
+int SetOption(int argc, char **argv, int *fft_size, float *q1,
     int *number_of_dimensions, char *filename) {
   while (--argc) {
     if (strcmp(argv[argc], "-f") == 0) *fft_size = atoi(argv[argc + 1]);
@@ -67,11 +67,11 @@ int SetOption(int argc, char **argv, int *fft_size, double *q1,
 }
 
 void WriteCodedSpectralEnvelope(char *filename,
-    const double * const *spectrogram, int fs, int f0_length,
-    double frame_period, int number_of_dimensions, int fft_size) {
-  double **coded_spectrogram = new double *[f0_length];
+    const float * const *spectrogram, int fs, int f0_length,
+    float frame_period, int number_of_dimensions, int fft_size) {
+  float **coded_spectrogram = new float *[f0_length];
   for (int i = 0; i < f0_length; ++i)
-    coded_spectrogram[i] = new double[number_of_dimensions];
+    coded_spectrogram[i] = new float[number_of_dimensions];
 
   CodeSpectralEnvelope(spectrogram, f0_length, fs, fft_size,
       number_of_dimensions, coded_spectrogram);
@@ -98,9 +98,9 @@ int main(int argc, char **argv) {
 
   // Read F0 information
   int f0_length = static_cast<int>(GetHeaderInformation(argv[2], "NOF "));
-  double frame_period = GetHeaderInformation(argv[2], "FP  ");
-  double *f0 = new double[f0_length];
-  double *temporal_positions = new double[f0_length];
+  float frame_period = GetHeaderInformation(argv[2], "FP  ");
+  float *f0 = new float[f0_length];
+  float *temporal_positions = new float[f0_length];
   ReadF0(argv[2], temporal_positions, f0);
 
   // Read an audio file
@@ -113,7 +113,7 @@ int main(int argc, char **argv) {
     }
     return -1;
   }
-  double *x = new double[x_length];
+  float *x = new float[x_length];
   int fs, nbit;
   wavread(argv[1], &fs, &nbit, x);
 
@@ -128,9 +128,9 @@ int main(int argc, char **argv) {
     &number_of_dimensions, filename) == 0) return -1;
 
   // Spectral envelope analysis
-  double **spectrogram = new double *[f0_length];
+  float **spectrogram = new float *[f0_length];
   for (int i = 0; i < f0_length; ++i)
-    spectrogram[i] = new double[option.fft_size / 2 + 1];
+    spectrogram[i] = new float[option.fft_size / 2 + 1];
   CheapTrick(x, x_length, fs, temporal_positions, f0, f0_length, &option,
       spectrogram);
